refactor: reorganize overlay classes
